use internal frame caching in DualFrameRawDecoder
We can't arbitrarily drop frames, because some sources (like the
WebcamInput) require all frames to be pulled (they cannot work out of
order).
So not pulling a frame deadlocks everything. So instead of skipping a
frame, we advance 1 frame instead of 2 in the DualFrameRawDecoder and
instead save the later of the two old frames for use in the next frame.