[WIP] arbitrary node graphs
This adds support for specifying a yaml based config file that allows
one to build arbitrary processing graphs.

To allow multiple sinks to run simultaneously, they need to be converted
to be actually async (atleast on a frame by frame basis).

Furthermore, switch from pollster to tokio. Currently this is does not
really bring us anything, but maybe we can replace some of the manual
thread handling with spawn_blocking later. Also we can now use
`tokio-console` :)

Limitations:
- Display currently is no longer running on the main thread,
which breaks on windows and macos. To fix this, we should add window
management and event reception parts to the ProcessingContext which
forwards them to a EventLoop running in the main thread.
- No caching is implemented, which makes running multiple sinks that
share parts of the processing chain really inefficient or impossible (in
case of nodes that only allow forward progress like Average or
DualFrameRawDecoder). To fix this we should finally implement the
caching node, for which some ground work is already done in #27.